+++ src/rvrbase/__init__.py
@@ -1,5 +1,5 @@
-from rvrbase.logger_api import * # noqa
-from rvrbase.config import * # noqa
-from rvrbase.rvrlogger import * # noqa
+from rvrbase.rvrlogger_api import * # noqa
+from rvrbase.rvrconfig_api import * # noqa
+from rvrbase.rvrlogger_delegate import * # noqa
 from rvrbase.constants import * # noqa
 from rvrbase.__about__ import (__version__,)
