+++ src/rvrbase/rvrlogger_api.py
@@ -9,14 +9,14 @@ import base64
 import logging
 import datetime
 from platform import node
-from . import rvrconfig
+import rvrbase
 from .constants import CONFIG_FILE, ERR_POST_DATA, LOG_TYPE_APPLICATION_EVENT, MSG_POST_DATA_SUCCESS
 
 
 class Log_analytics_logger:
 
     def __init__(self):
-        self.conf = rvrconfig.Rvrconfig(CONFIG_FILE)
+        self.conf = rvrbase.Rvrconfig(CONFIG_FILE)
 
     def build_signature(self, customer_id, shared_key, date, content_length, method, content_type,
                         resource):
